app/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-session==0.6.0
flask-mail==0.9.1
flask-socketio==5.3.6
flask-bcrypt==1.0.1
flask-limiter==3.5.0
flask-caching==2.1.0
python-dotenv==1.0.0
itsdangerous==2.1.2
scikit-learn==1.4.0
numpy==1.26.4
Pillow==10.2.0
gunicorn==21.2.0
eventlet==0.35.1
sqlalchemy==2.0.25